added first test
