+++ backend/src/routers/custom_tiles/sql_constants.py
@@ -1,3 +1,13 @@
+# Connection Constants
+DB_URL_ENV_VAR="CT_DB_URL"
+DB_PORT_ENV_VAR="CT_DB_PORT"
+DB_USERNAME_ENV_VAR="CT_DB_USERNAME"
+DB_PASSWORD_ENV_VAR="CT_DB_PASSWORD"
+
+
+# Error constants
+DB_CONNECT_FAILURE_MSG="DB_FAILED_TO_CONNECT"
+
 INSERT_CUSTOM_TILE_QUERY = (
 	"INSERT INTO SmartSpeechCustomTiles.CustomTiles "
 	"(ImageURL, UserEmail, TextAssociated, SoundAssociated, TileColor) "
